Module eclair::alloc

Compiler Allocation

When building programs, we may need to transfer values across compilers, say from a native compiler into a circuit-building compiler. This lifting of values into another compiler is known as “allocation”. The two traits responsible for this allocation scheme are Constant and Variable. See their documentation for more.

Note on Terminology

Allocation does not necessarily refer to allocation of memory nor the simulation of memory-based abstractions inside of compilers, like heap allocation. Allocation only refers to lifting constants and variables from one compiler to another.

Modules

mode

Allocation Modes for Variable Types

Traits

Allocate

Allocation Auto-trait

Allocator

Allocator Auto-trait

Constant

Compiler Constant

Variable

Compiler Variable

Type Definitions

Const

Constant Type Alias

Var

Variable Type Alias