Trait eclair::alloc::Constant

pub trait Constant<COM = ()>where
    COM: ?Sized,{
    type Type;

    fn new_constant(this: &Self::Type, compiler: &mut COM) -> Self;
}

Compiler Constant

A compiler Constant is a kind of allocated value that must be known at compilation time. In this case, we take an explicit value of the underlying type and directly return an allocated value in the COM compiler. Contrast this with Variable types whose values are not known at compilation time and are only known at execution time.

Required Associated Types

type Type

Underlying Type

Required Methods

fn new_constant(this: &Self::Type, compiler: &mut COM) -> Self

Allocates a new constant from this into the compiler.