Trait eclair::alloc::Variable

pub trait Variable<M, COM = ()>where
    COM: ?Sized,{
    type Type;

    fn new_unknown(compiler: &mut COM) -> Self;
    fn new_known(this: &Self::Type, compiler: &mut COM) -> Self;
}

Compiler Variable

A compiler Variable is a kind of allocated value that cannot have a known value at compilation time but does have one at execution time. In this case, we require two allocation methods new_unknown and new_known in order to use the variables in COM during both the compilation and execution phases. Contrast this with Constant types whose values must be known at compilation time.

Allocation Modes

The tag M in the type parameters of this trait refers to the Variable’s allocation mode. See the mode module for more details on why allocation modes are necessary and useful and how to use them.

Required Associated Types

type Type

Underlying Type

Required Methods

fn new_unknown(compiler: &mut COM) -> Self

Allocates a new unknown value into the compiler. The terminology “unknown” refers to the fact that we need to allocate a slot for this variable during compilation time, but do not yet know its underlying value.

fn new_known(this: &Self::Type, compiler: &mut COM) -> Self

Allocates a new known value from this into the compiler. The terminology “known” refers to the fact that we have access to the underyling value during execution time where we are able to use its concrete value for execution.