Trait eclair::bool::ConditionalSelect

pub trait ConditionalSelect<COM = ()>: Sizedwhere
    COM: Has<bool> + ?Sized,{
    fn select(
        bit: &Bool<COM>,
        true_value: &Self,
        false_value: &Self,
        compiler: &mut COM
    ) -> Self;

    fn select_from_table<'s, B, T>(bits: B, table: T, compiler: &mut COM) -> Self
    where
        Self: 's + Clone,
        Bool<COM>: 's,
        B: IntoIterator<Item = &'s Bool<COM>>,
        B::IntoIter: ExactSizeIterator,
        T: IntoIterator<Item = &'s Self>,
        T::IntoIter: ExactSizeIterator,
    { ... }
}

Conditional Selection

Required Methods

fn select(
bit: &Bool<COM>,
true_value: &Self,
false_value: &Self,
compiler: &mut COM
) -> Self

Selects true_value when bit == true and false_value when bit == false.

Provided Methods

fn select_from_table<'s, B, T>(bits: B, table: T, compiler: &mut COM) -> Selfwhere
Self: 's + Clone,
Bool<COM>: 's,
B: IntoIterator<Item = &'s Bool<COM>>,
B::IntoIter: ExactSizeIterator,
T: IntoIterator<Item = &'s Self>,
T::IntoIter: ExactSizeIterator,

Available on crate feature alloc only.

Selects an element from table by repeated iteration of select over bits. The bits are ordered from most significant to least significant, forming unsigned integers in binary representation which are understood as the table indices.