Trait eclair::cmp::PartialEq

pub trait PartialEq<Rhs, COM = ()>where
    Rhs: ?Sized,
    COM: Has<bool> + ?Sized,{
    fn eq(&self, rhs: &Rhs, compiler: &mut COM) -> Bool<COM>;

    fn ne(&self, other: &Rhs, compiler: &mut COM) -> Bool<COM>
    where
        Bool<COM>: Not<COM, Output = Bool<COM>>,
    { ... }
    fn assert_equal(&self, rhs: &Rhs, compiler: &mut COM)
    where
        COM: Assert,
    { ... }
}

Partial Equivalence Relations

Required Methods

fn eq(&self, rhs: &Rhs, compiler: &mut COM) -> Bool<COM>

Returns true if self and rhs are equal.

Provided Methods

fn ne(&self, other: &Rhs, compiler: &mut COM) -> Bool<COM>where
Bool<COM>: Not<COM, Output = Bool<COM>>,

Returns true if self and rhs are not equal.

fn assert_equal(&self, rhs: &Rhs, compiler: &mut COM)where
COM: Assert,

Asserts that self and rhs are equal.

Implementation Note

This method is an optimization path for the case when comparing for equality and then asserting is more expensive than a custom assertion.